Implementirano uređivanje recepta (naziv, vrijeme, opis, kategorija i preferencija)
- dodana funckionalnost za uređivanje recepata (za sad se može uređivati naziv, vrijeme izrade, opis recepta, njegova kategorija i preferencija)
- ažurirano polje Quantity (u Ingredient) u double
- izmjenjena aktivnost EditRecipeActivity

+++ Software/app/src/main/java/com/example/receptomat/entities/Ingredient.kt
@@ -2,7 +2,7 @@ package com.example.receptomat.entities
 
 data class Ingredient(
     val item_name: String,
-    val quantity: Float,
+    val quantity: Double,
     val unit_id: Int,
     val unit_name: String?
 )
